pyproject.toml
[project]
name = "geodatasets"
version = "0.1.0"
description = "Package for conversion of geo datasets to different formats."
authors = [
    {name = "Miriam Strihavkova"}
]
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32",
    "tqdm>=4.67",
    "numpy>=2.4.3,<3.0.0",
    "pandas>=3.0.1,<4.0.0",
    "plotly>=6.6.0,<7.0.0",
    "pillow>=12.1.1,<13.0.0",
    "nbformat>=5.10.4,<6.0.0",
    "rasterio>=1.3",
    "h5py>=3.9",
    "pydantic>=2.0",
    "torch>=2.0",
    "pytorch-lightning>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
geodatasets-download = "geodatasets.scripts.download:main"
geodatasets-process-s2  = "geodatasets.scripts.process_s2_catalogue:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.8,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "types-requests (>=2.33.0.20260327,<3.0.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "pydocstyle (>=6.3.0,<7.0.0)"
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.mypy]
ignore_missing_imports = true

[tool.pydocstyle]
convention = "google"
add-ignore = ["D107", "D105"]
match = "(?!test_).*\\.py"